refactor(onboard): extract portable bootstrap guard
